Backend fixes

+++ App/Routes/PatientRoute.py
@@ -2,21 +2,22 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from ..Models.PatientModel import Patient
 from sqlalchemy.orm import Session
 from App.Data.Database import get_db
-from ..Repos.PatientRepo import PatientsRepo
+from ..Repos.PatientRepo import PatientRepo
 
 router = APIRouter(
-  prefix="/patients"
+  prefix="/patients",
+  tags=["Patients"]
 )
 
 @router.get("")
 async def get_patients(db: Session = Depends(get_db)):
-  Patients = await PatientsRepo(db).get_patients()
-  return Patients
+  repo = PatientRepo(db)
+  patients = await repo.get_patients()
+  return patients
 
 @router.get("/{id}")
 async def get_patientById(id: int, db: Session = Depends(get_db)):
-  repo = PatientsRepo(db)
-
+  repo = PatientRepo(db)
   exists = await repo.patientExists(id)
   if exists < 1:
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -32,7 +33,7 @@ async def add_patient(patient: Patient, db: Session = Depends(get_db)):
 
 @router.put("/{id}")
 async def update_patient(id:int, patient: Patient, db: Session = Depends(get_db)):
-  repo = PatientsRepo(db)
+  repo = PatientRepo(db)
   exists = await repo.patientExists(id)
   
   if exists < 1:
@@ -40,15 +41,16 @@ async def update_patient(id:int, patient: Patient, db: Session = Depends(get_db)
                         detail=f"Patient with id:{id} not found or doesn't exist!")
   
   await repo.update_patient(id, patient)
-  return await repo.get_patient(id)
-    
+  return f"Patient with id:{id} updated!"
+
 @router.delete("/{id}")
 async def delete_patient(id: int, db: Session = Depends(get_db)):
-  repo = PatientsRepo(db)
-  isDeleted = await repo.delete_patient(id)
-
-  if isDeleted:
-     return f"Patient with id:{id} has been succesfuly removed!"
+  repo = PatientRepo(db)
+  exists = await repo.patientExists(id)
   
-  raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+  if exists < 1:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                         detail=f"Patient with id:{id} not found or doesn't exist!")
+  
+  await repo.delete_patient(id)
+  return f"Patient with id:{id} deleted!"
